use float instead of double

+++ libft/ft_atod.c
@@ -3,24 +3,24 @@
 /*                                                        :::      ::::::::   */
 /*   ft_atod.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: aogbi <aogbi@student.42.fr>                +#+  +:+       +#+        */
+/*   By: aogbi <aogbi@student.1337.ma>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/01 23:44:43 by aogbi             #+#    #+#             */
-/*   Updated: 2025/01/03 10:40:14 by aogbi            ###   ########.fr       */
+/*   Updated: 2025/02/10 19:16:25 by aogbi            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-double	ft_atod(char *str)
+float	ft_atod(char *str)
 {
-	double	result;
-	double	decimal;
+	float	result;
+	float	decimal;
 	int		count;
 	int		sign;
 
-	result = 0;
-	decimal = 0;
+	result = 0.0f;
+	decimal = 0.0f;
 	sign = 1;
 	if (*str == '-')
 	{
@@ -28,17 +28,17 @@ double	ft_atod(char *str)
 		str++;
 	}
 	while (ft_isdigit(*str))
-		result = result * 10 + (*str++ - '0');
+		result = result * 10.0f + (float)(*str++ - '0');
 	count = 0;
 	if (*str == '.')
 	{
 		str++;
 		while (ft_isdigit(*str))
 		{
-			decimal = decimal * 10 + (*str - '0');
+			decimal = decimal * 10.0f + (float)(*str - '0');
 			str++;
 			count++;
 		}
 	}
-	return ((result + decimal / pow(10, count)) * sign);
+	return ((result + decimal / (float)pow(10, count)) * (float)sign);
 }
